dex: break out engine idx into modules
