+++ crates/core/component/dex/src/component/position_manager.rs
@@ -8,23 +8,28 @@ use cnidarium::{EscapedByteSlice, StateRead, StateWrite};
 use futures::Stream;
 use futures::StreamExt;
 use penumbra_asset::{asset, Balance};
-use penumbra_num::Amount;
 use penumbra_proto::DomainType;
 use penumbra_proto::{StateReadProto, StateWriteProto};
 
-use crate::lp::position::State;
+use crate::component::position_manager::{
+    base_liquidity_index::AssetByLiquidityIndex, inventory_index::PositionByInventoryIndex,
+    price_index::PositionByPriceIndex,
+};
 use crate::lp::Reserves;
 use crate::{
     component::position_counter::PositionCounter,
     component::ValueCircuitBreaker,
     lp::position::{self, Position},
     state_key::engine,
-    state_key::eviction_queue,
     DirectedTradingPair,
 };
 use crate::{event, state_key};
 
 const DYNAMIC_ASSET_LIMIT: usize = 10;
+
+pub(crate) mod base_liquidity_index;
+pub(crate) mod inventory_index;
+pub(crate) mod price_index;
 
 #[async_trait]
 pub trait PositionRead: StateRead {
@@ -429,258 +434,6 @@ pub(crate) trait Inner: StateWrite {
             .await?;
 
         self.put(state_key::position_by_id(&id), new_state);
-        Ok(())
-    }
-
-    // TODO(erwan): break this out into a `position_manager::inventory_index` module.
-    fn index_position_by_inventory(&mut self, position: &position::Position, id: &position::Id) {
-        tracing::debug!("indexing position by inventory");
-        let canonical_pair = position.phi.pair;
-        // A position is bound to an unordered trading pair: A <> B.
-        // We want to index the position by inventory for each direction:
-        // A -> B
-        let pair_ab = DirectedTradingPair::new(canonical_pair.asset_1(), canonical_pair.asset_2());
-        let inventory_a = position
-            .reserves_for(pair_ab.start)
-            .expect("the directed trading pair is correct");
-        let key_ab = eviction_queue::inventory_index::key(&pair_ab, inventory_a, id).to_vec();
-        self.nonverifiable_put_raw(key_ab, vec![]);
-
-        // B -> A
-        let pair_ba = pair_ab.flip();
-        let inventory_b = position
-            .reserves_for(pair_ba.start)
-            .expect("the directed trading pair is correct");
-        let key_ba = eviction_queue::inventory_index::key(&pair_ba, inventory_b, id).to_vec();
-        self.nonverifiable_put_raw(key_ba, vec![]);
-    }
-
-    fn deindex_position_by_inventory(
-        &mut self,
-        prev_position: &position::Position,
-        id: &position::Id,
-    ) {
-        let canonical_pair = prev_position.phi.pair;
-
-        // To deindex the position, we need to reconstruct the tuple of keys
-        // that correspond to each direction of the trading pair:
-        // A -> B
-        let pair_ab = DirectedTradingPair::new(canonical_pair.asset_1(), canonical_pair.asset_2());
-        let inventory_a = prev_position
-            .reserves_for(pair_ab.start)
-            .expect("the directed trading pair is correct");
-        let key_ab = eviction_queue::inventory_index::key(&pair_ab, inventory_a, id).to_vec();
-        self.nonverifiable_delete(key_ab);
-
-        // B -> A
-        let pair_ba = pair_ab.flip();
-        let inventory_b = prev_position
-            .reserves_for(pair_ba.start)
-            .expect("the directed trading pair is correct");
-        let key_ba = eviction_queue::inventory_index::key(&pair_ba, inventory_b, id).to_vec();
-        self.nonverifiable_delete(key_ba);
-    }
-
-    fn index_position_by_price(&mut self, position: &position::Position, id: &position::Id) {
-        let (pair, phi) = (position.phi.pair, &position.phi);
-        if position.reserves.r2 != 0u64.into() {
-            // Index this position for trades FROM asset 1 TO asset 2, since the position has asset 2 to give out.
-            let pair12 = DirectedTradingPair {
-                start: pair.asset_1(),
-                end: pair.asset_2(),
-            };
-            let phi12 = phi.component.clone();
-            self.nonverifiable_put_raw(engine::price_index::key(&pair12, &phi12, &id), vec![]);
-            tracing::debug!("indexing position for 1=>2 trades");
-        }
-
-        if position.reserves.r1 != 0u64.into() {
-            // Index this position for trades FROM asset 2 TO asset 1, since the position has asset 1 to give out.
-            let pair21 = DirectedTradingPair {
-                start: pair.asset_2(),
-                end: pair.asset_1(),
-            };
-            let phi21 = phi.component.flip();
-            self.nonverifiable_put_raw(engine::price_index::key(&pair21, &phi21, &id), vec![]);
-            tracing::debug!("indexing position for 2=>1 trades");
-        }
-    }
-
-    fn deindex_position_by_price(&mut self, position: &Position, id: &position::Id) {
-        tracing::debug!("deindexing position");
-        let pair12 = DirectedTradingPair {
-            start: position.phi.pair.asset_1(),
-            end: position.phi.pair.asset_2(),
-        };
-        let phi12 = position.phi.component.clone();
-        let pair21 = DirectedTradingPair {
-            start: position.phi.pair.asset_2(),
-            end: position.phi.pair.asset_1(),
-        };
-        let phi21 = position.phi.component.flip();
-        self.nonverifiable_delete(engine::price_index::key(&pair12, &phi12, &id));
-        self.nonverifiable_delete(engine::price_index::key(&pair21, &phi21, &id));
-    }
-
-    /// Updates the nonverifiable liquidity indices given a [`Position`] in the direction specified by the [`DirectedTradingPair`].
-    /// An [`Option<Position>`] may be specified to allow for the case where a position is being updated.
-    async fn update_liquidity_index(
-        &mut self,
-        pair: DirectedTradingPair,
-        position: &Position,
-        prev: &Option<Position>,
-    ) -> Result<()> {
-        tracing::debug!(?pair, "updating available liquidity indices");
-
-        let (new_a_from_b, current_a_from_b) = match (position.state, prev) {
-            (State::Opened, None) => {
-                // Add the new position's contribution to the index, no cancellation of the previous version necessary.
-
-                // Query the current available liquidity for this trading pair, or zero if the trading pair
-                // has no current liquidity.
-                let current_a_from_b = self
-                    .nonverifiable_get_raw(&engine::routable_assets::a_from_b(&pair))
-                    .await?
-                    .map(|bytes| {
-                        Amount::from_be_bytes(
-                            bytes
-                                .try_into()
-                                .expect("liquidity index amount can always be parsed"),
-                        )
-                    })
-                    .unwrap_or_default();
-
-                // Use the new reserves to compute `new_position_contribution`,
-                // the amount of asset A contributed by the position (i.e. the reserves of asset A).
-                let new_position_contribution = position
-                    .reserves_for(pair.start)
-                    .expect("specified position should match provided trading pair");
-
-                // Compute `new_A_from_B`.
-                let new_a_from_b =
-                    // Add the contribution from the updated version.
-                    current_a_from_b.saturating_add(&new_position_contribution);
-
-                tracing::debug!(?pair, current_liquidity = ?current_a_from_b, ?new_position_contribution, "newly opened position, adding contribution to existing available liquidity for trading pair");
-
-                (new_a_from_b, current_a_from_b)
-            }
-            (State::Opened, Some(prev)) => {
-                // Add the new position's contribution to the index, deleting the previous version's contribution.
-
-                // Query the current available liquidity for this trading pair, or zero if the trading pair
-                // has no current liquidity.
-                let current_a_from_b = self
-                    .nonverifiable_get_raw(&engine::routable_assets::a_from_b(&pair))
-                    .await?
-                    .map(|bytes| {
-                        Amount::from_be_bytes(
-                            bytes
-                                .try_into()
-                                .expect("liquidity index amount can always be parsed"),
-                        )
-                    })
-                    .unwrap_or_default();
-
-                // Use the previous reserves to compute `prev_position_contribution` (denominated in asset_1).
-                let prev_position_contribution = prev
-                    .reserves_for(pair.start)
-                    .expect("specified position should match provided trading pair");
-
-                // Use the new reserves to compute `new_position_contribution`,
-                // the amount of asset A contributed by the position (i.e. the reserves of asset A).
-                let new_position_contribution = position
-                    .reserves_for(pair.start)
-                    .expect("specified position should match provided trading pair");
-
-                // Compute `new_A_from_B`.
-                let new_a_from_b =
-                // Subtract the previous version of the position's contribution to represent that position no longer
-                // being correct, and add the contribution from the updated version.
-                (current_a_from_b.saturating_sub(&prev_position_contribution)).saturating_add(&new_position_contribution);
-
-                tracing::debug!(?pair, current_liquidity = ?current_a_from_b, ?new_position_contribution, ?prev_position_contribution, "updated position, adding new contribution and subtracting previous contribution to existing available liquidity for trading pair");
-
-                (new_a_from_b, current_a_from_b)
-            }
-            (State::Closed, Some(prev)) => {
-                // Compute the previous contribution and erase it from the current index
-
-                // Query the current available liquidity for this trading pair, or zero if the trading pair
-                // has no current liquidity.
-                let current_a_from_b = self
-                    .nonverifiable_get_raw(&engine::routable_assets::a_from_b(&pair))
-                    .await?
-                    .map(|bytes| {
-                        Amount::from_be_bytes(
-                            bytes
-                                .try_into()
-                                .expect("liquidity index amount can always be parsed"),
-                        )
-                    })
-                    .unwrap_or_default();
-
-                // Use the previous reserves to compute `prev_position_contribution` (denominated in asset_1).
-                let prev_position_contribution = prev
-                    .reserves_for(pair.start)
-                    .expect("specified position should match provided trading pair");
-
-                // Compute `new_A_from_B`.
-                let new_a_from_b =
-                // Subtract the previous version of the position's contribution to represent that position no longer
-                // being correct, and since the updated version is Closed, it has no contribution.
-                current_a_from_b.saturating_sub(&prev_position_contribution);
-
-                tracing::debug!(?pair, current_liquidity = ?current_a_from_b, ?prev_position_contribution, "closed position, subtracting previous contribution to existing available liquidity for trading pair");
-
-                (new_a_from_b, current_a_from_b)
-            }
-            (State::Withdrawn { .. }, _) | (State::Closed, None) => {
-                // The position already went through the `Closed` state or was opened in the `Closed` state, so its contribution has already been subtracted.
-                return Ok(());
-            }
-        };
-
-        // Delete the existing key for this position if the reserve amount has changed.
-        if new_a_from_b != current_a_from_b {
-            self.nonverifiable_delete(
-                engine::routable_assets::key(&pair.start, current_a_from_b).to_vec(),
-            );
-        }
-
-        // Write the new key indicating that asset B is routable from asset A with `new_a_from_b` liquidity.
-        self.nonverifiable_put_raw(
-            engine::routable_assets::key(&pair.start, new_a_from_b).to_vec(),
-            pair.end.encode_to_vec(),
-        );
-        tracing::debug!(start = ?pair.start, end = ?pair.end, "marking routable from start -> end");
-
-        // Write the new lookup index storing `new_a_from_b` for this trading pair.
-        self.nonverifiable_put_raw(
-            engine::routable_assets::a_from_b(&pair).to_vec(),
-            new_a_from_b.to_be_bytes().to_vec(),
-        );
-        tracing::debug!(available_liquidity = ?new_a_from_b, ?pair, "marking available liquidity for trading pair");
-
-        Ok(())
-    }
-
-    async fn update_available_liquidity(
-        &mut self,
-        prev_position: &Option<Position>,
-        position: &Position,
-    ) -> Result<()> {
-        // Since swaps may be performed in either direction, the available liquidity indices
-        // need to be calculated and stored for both the A -> B and B -> A directions.
-        let (a, b) = (position.phi.pair.asset_1(), position.phi.pair.asset_2());
-
-        // A -> B
-        self.update_liquidity_index(DirectedTradingPair::new(a, b), position, prev_position)
-            .await?;
-        // B -> A
-        self.update_liquidity_index(DirectedTradingPair::new(b, a), position, prev_position)
-            .await?;
-
         Ok(())
     }
 }
